refactor: rename accordion to collapsible
